backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
mysqlclient>=2.2,<2.3
Pillow>=10.4,<11.0

